CLN: move color plot options back onto Scatter plot class
