+++ force_wfmanager/ui/review/scatter_plot.py
@@ -7,10 +7,23 @@
 
 """
 
+from chaco.default_colormaps import color_map_name_dict
+from chaco.api import Plot as ChacoPlot
 from chaco.api import ScatterInspectorOverlay
 from chaco.tools.api import BetterSelectingZoom as ZoomTool
 from chaco.tools.api import PanTool, ScatterInspector
 from enable.api import KeySpec
+from traits.api import (
+    Button,
+    Bool,
+    Dict,
+    Enum,
+    List,
+    on_trait_change,
+    Str,
+    Property
+)
+from traitsui.api import EnumEditor, HGroup, Item, UItem, View
 
 from .base_plot import BasePlot
 
@@ -20,6 +33,113 @@ class ScatterPlot(BasePlot):
 
     #: Short description for the UI selection
     description = "Scatter plot with colormap"
+
+    # ------------------
+    # Regular Attributes
+    # ------------------
+
+    #: Whether or not to include individually colored data points
+    #: in displayed plot
+    use_color_plot = Bool(False)
+
+    #: Optional third parameter used to set colour of points
+    color_by = Enum(values="displayable_value_names")
+
+    #: Colour options button:
+    color_options = Button("Color...")
+
+    #: Available color maps provided by plotting backend
+    _available_color_maps = Dict(color_map_name_dict)
+
+    # --------------------
+    # Dependent Attributes
+    # --------------------
+
+    #: Selectable color maps
+    colormap = Enum(
+        values="_available_color_map_names",
+        depends_on="_available_color_map_names",
+    )
+
+    # ----------
+    # Properties
+    # ----------
+
+    #: Available color maps names provided by plotting backend
+    _available_color_map_names = Property(
+        List(Str), depends_on='_available_color_maps')
+
+    # ----
+    # View
+    # ----
+
+    def _color_options_fired(self):
+        """ Event handler for :attr:`color_options` button. """
+        view = View(
+            Item("use_color_plot"),
+            Item("color_by", enabled_when="use_color_plot"),
+            Item("colormap", enabled_when="use_color_plot"),
+            kind="livemodal",
+        )
+        self.edit_traits(view=view)
+
+    def _axis_hgroup_default(self):
+        return HGroup(
+            Item("x", editor=EnumEditor(name="displayable_value_names")),
+            Item("y", editor=EnumEditor(name="displayable_value_names")),
+            UItem("color_options"),
+            Item("toggle_automatic_update", label="Axis auto update"),
+        )
+
+    # ----------
+    # Properties
+    # ----------
+
+    def _get__available_color_map_names(self):
+        return ["viridis"] + [
+            cmap_name
+            for cmap_name in self._available_color_maps.keys()
+            if cmap_name != "viridis"
+        ]
+
+    # ---------
+    # Listeners
+    # ---------
+
+    @on_trait_change("color_by")
+    def _update_color_plot(self):
+        if (
+            self.x == ""
+            or self.y == ""
+            or self.color_by is None
+            or self.analysis_model.is_empty
+        ):
+            self._plot_data.set_data("color_by", [])
+            return
+
+        self._plot_data.set_data(
+            "color_by", self.analysis_model.column(self.color_by)
+        )
+
+    @on_trait_change("colormap")
+    def _update_cmap(self):
+        cmap = self._available_color_maps[self.colormap]
+        if hasattr(self._axis, 'color_mapper'):
+            _range = self._axis.color_mapper.range
+            self._axis.color_mapper = cmap(_range)
+
+    @on_trait_change("use_color_plot")
+    def change_plot_style(self):
+        ranges = self._get_plot_range()
+        x_title = self._plot.x_axis.title
+        y_title = self._plot.y_axis.title
+
+        self._plot = ChacoPlot(self._plot_data)
+        self.add_plots(self._plot)
+
+        self._set_plot_range(*ranges)
+        self._plot.x_axis.title = x_title
+        self._plot.y_axis.title = y_title
 
     def _add_plot_tools(self, plot):
         """Add pan and zoom tools"""
@@ -97,3 +217,9 @@ class ScatterPlot(BasePlot):
 
     def add_plots(self, plot):
         self.plot_scatter(plot)
+
+    def add_plot_data(self, plot_data):
+        plot_data.set_data("color_by", [])
+
+    def update_data_view(self):
+        self._update_color_plot()
