+++ force_wfmanager/ui/review/tests/test_scatter_plot.py
@@ -7,7 +7,7 @@ from chaco.api import Plot as ChacoPlot
 from chaco.api import ScatterPlot as ChacoScatterPlot
 from chaco.api import ColormappedScatterPlot
 from chaco.abstract_colormap import AbstractColormap
-from traits.api import push_exception_handler
+from traits.api import push_exception_handler, TraitError
 
 from force_wfmanager.ui.review.scatter_plot import ScatterPlot
 
@@ -19,6 +19,15 @@ push_exception_handler(reraise_exceptions=True)
 class TestScatterPlot(BasePlotTestCase):
 
     plot_cls = ScatterPlot
+
+    def test_init(self):
+        self.assertEqual("viridis", self.plot._available_color_map_names[0])
+
+        self.plot._update_plot()
+        self.assertEqual(
+            [],
+            self.plot._plot_data.get_data("color_by").tolist()
+        )
 
     def test_plot(self):
         self.analysis_model.header = ("density", "pressure")
@@ -52,7 +61,21 @@ class TestScatterPlot(BasePlotTestCase):
                 old_cmap.range, self.plot._axis.color_mapper.range
             )
 
+        with self.assertRaises(TraitError):
+            self.plot.colormap = "not_viridis"
+
         self.plot.colormap = "viridis"
         self.plot.colormap = "CoolWarm"
         self.plot.use_color_plot = False
         self.assertIsInstance(self.plot._axis, ChacoScatterPlot)
+
+    def test_ranges_are_kept(self):
+        self.analysis_model.header = ("density", "pressure", "color")
+        self.analysis_model.notify((1.010, 101325, 1))
+        self.check_update_is_requested_and_apply()
+        self.plot._set_plot_range(0.5, 2, 100000, 103000)
+        self.plot.use_color_plot = True
+        self.plot.color_by = "color"
+        self.assertEqual(self.plot._get_plot_range(), (0.5, 2, 100000, 103000))
+        self.assertEqual(self.plot._plot.x_axis.title, "density")
+        self.assertEqual(self.plot._plot.y_axis.title, "pressure")
